pyproject.toml
[project]
name = "nlp-projects"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.13.0",
    "datasets>=4.8.4",
    "gemini-cli>=0.2.5",
    "numpy>=2.4.4",
    "opencv-python>=4.13.0.92",
    "pandas>=3.0.2",
    "pillow>=12.2.0",
    "scikit-learn>=1.8.0",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
    "transformers>=5.5.4",
]
